Create framework around k6 testing tool
