+++ package.json
@@ -4,8 +4,8 @@
   "description": "Load testing scripts for Pelias",
   "main": "index.js",
   "dependencies": {
-    "byline": "^5.0.0",
-    "loadtest": "^2.2.1"
+    "csv-parse": "^2.0.0",
+    "lodash": "^4.17.4"
   },
   "devDependencies": {},
   "scripts": {
